Move ignore_layer & max2avg_pool2d to utils/models
