+++ captum/optim/_utils/models.py
@@ -1,12 +1,12 @@
 import math
-from typing import List, Tuple, Union
+from typing import Any, List, Optional, Tuple, Union
 
 import torch
 import torch.nn as nn
 import torch.nn.functional as F
 
 from captum.optim._core.output_hook import ActivationFetcher
-from captum.optim._utils.typing import ModelInputType, ModuleOutputMapping
+from captum.optim._utils.typing import ModelInputType, ModuleOutputMapping, PoolParam
 
 
 def get_model_layers(model) -> List[str]:
@@ -172,3 +172,75 @@ def collect_activations(
     catch_activ = ActivationFetcher(model, targets)
     activ_out = catch_activ(model_input)
     return activ_out
+
+
+def max2avg_pool2d(model, value: Optional[Any] = float("-inf")) -> None:
+    """
+    Replace all nonlinear MaxPool2d layers with their linear AvgPool2d equivalents.
+    This allows us to ignore nonlinear values when calculating expanded weights.
+
+    Args:
+        model (nn.Module): A PyTorch model instance.
+        value (Any): Used to return any padding that's meant to be ignored by
+            pooling layers back to zero.
+    """
+
+    class AvgPool2dLayer(torch.nn.Module):
+        def __init__(
+            self,
+            kernel_size: PoolParam = 2,
+            stride: Optional[PoolParam] = 2,
+            padding: PoolParam = 0,
+            ceil_mode: bool = False,
+            value: Optional[Any] = None,
+        ) -> None:
+            super().__init__()
+            self.avgpool = torch.nn.AvgPool2d(
+                kernel_size=kernel_size,
+                stride=stride,
+                padding=padding,
+                ceil_mode=ceil_mode,
+            )
+            self.value = value
+
+        def forward(self, x: torch.Tensor) -> torch.Tensor:
+            x = self.avgpool(x)
+            if self.value is not None:
+                x[x == self.value] = 0.0
+            return x
+
+    for name, child in model._modules.items():
+        if isinstance(child, torch.nn.MaxPool2d):
+            new_layer = AvgPool2dLayer(
+                kernel_size=child.kernel_size,
+                stride=child.stride,
+                padding=child.padding,
+                ceil_mode=child.ceil_mode,
+                value=value,
+            )
+            setattr(model, name, new_layer)
+        elif child is not None:
+            max2avg_pool2d(child)
+
+
+def ignore_layer(model, layer) -> None:
+    """
+    Replace target layers with layers that do nothing.
+    This is useful for removing the nonlinear ReLU
+    layers when creating expanded weights.
+
+    Args:
+        model (nn.Module): A PyTorch model instance.
+        layer (nn.Module): A layer class type.
+    """
+
+    class IgnoreLayer(torch.nn.Module):
+        def forward(self, x: torch.Tensor) -> torch.Tensor:
+            return x
+
+    for name, child in model._modules.items():
+        if isinstance(child, layer):
+            new_layer = IgnoreLayer()
+            setattr(model, name, new_layer)
+        elif child is not None:
+            ignore_layer(child, layer)
